# server/app/generated/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2024-11-07T19:06:13+00:00

from __future__ import annotations

from typing import List, Optional, Union

from fastapi import FastAPI, UploadFile

from .models import (
    TranscribersGetResponseItem,
    TranscriptionsGetResponseItem,
    TranscriptionsPostResponse,
)

app = FastAPI(
    title='Audio Transcription API',
    description='API for transcribing audio files using various transcription engines',
    version='1.0.0',
)


@app.get('/transcribers', response_model=List[TranscribersGetResponseItem])
def get_transcribers() -> List[TranscribersGetResponseItem]:
    """
    List available transcribers
    """
    pass


@app.post(
    '/transcriptions',
    response_model=None,
    responses={'201': {'model': TranscriptionsPostResponse}},
)
def post_transcriptions(file: UploadFile) -> Optional[TranscriptionsPostResponse]:
    """
    Transcribe an audio file
    """
    pass


@app.get('/transcriptions', response_model=List[TranscriptionsGetResponseItem])
def get_transcriptions() -> List[TranscriptionsGetResponseItem]:
    """
    List past transcriptions
    """
    pass


@app.delete('/transcriptions/{transcription_id}', response_model=None)
def delete_transcriptions_transcription_id(transcription_id: str) -> None:
    """
    Delete a transcription
    """
    pass